Changed: Corrected the problem of favorite city dates from db showing wrong Changed: Now removing favorite also removes from db

+++ app/src/main/java/com/example/akarsh/hw06/FavoriteCityTableDAO.java
@@ -90,7 +90,7 @@ public class FavoriteCityTableDAO {
         } else {
             favoriteCity.setFavorite(false);
         }
-        Date updatedDate = new Date(c.getInt(4));
+        Date updatedDate = new Date(c.getLong(4));
         favoriteCity.setUpdated(updatedDate);
         return favoriteCity;
     }
